Squashed commit of the following:
commit 24c75a9dc69c3b57092f10e87f616ed4e2e106a9
Author: Ratchet <250710285+RatchetStuff@users.noreply.github.com>
Date:   Sun Sep 20 19:46:09 2026 +0800

    feat(prepare.yml): Extend sleep time

commit f98c367b88376adcc8534e8d245f191946c3045a
Author: Ratchet <250710285+RatchetStuff@users.noreply.github.com>
Date:   Sun Sep 20 19:37:56 2026 +0800

    chore: Cleanup workflows and shell scripts

+++ export.sh
@@ -1,7 +1,10 @@
 TOPLEVEL=$(git rev-parse --show-toplevel) || exit 1
 BIN="$TOPLEVEL/bin"
-rm -r "$BIN"
+if [ -d "$BIN" ]; then
+    rm -r "$BIN"
+fi
 mkdir -p "$BIN"
+
 find "$TOPLEVEL/versions" -type f -name "pack.toml" -print0 | while IFS= read -r -d '' filepath; do
     dir=$(dirname "$filepath")
     distribution="$(basename "$dir")"
